chore: add naming into stage2
Most of our coding time is spent reading existing code. Therefore, it should be as readable and understandable as possible for us humans. We do this by choosing words and constructing sentences using our code that are as close as possible to the human language we naturally use. Naming is a key part of programming skills.